pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "new_ltpp"
version = "0.1.0"
description = "An easy and flexible tool for neural temporal point process"
readme = "README.md"
authors = [{ name = "Alipay" }]
license = { text = "MIT" }
requires-python = "==3.11.*"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["temporal point process", "neural networks", "machine learning"]
dependencies = [
    "numpy>=2.3.3",
    "pandas>=2.0.0",
    "torch>=2.0.0",
    "lightning>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.16.2",
    "wandb>=0.22.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "torchmetrics>=1.0.0",
    "tensorboard>=2.15.0",
    "tqdm>=4.65.0",
    "fastdtw>=0.3.4",
    "datasets>=4.1.1",
    "polars>=1.33.1",
    "optuna>=4.5.0",
    "ipykernel>=6.30.1",
    "typer>=0.19.2",
    "pytest>=8.4.2",
    "pip>=25.3",
    "git-filter-repo>=2.47.0",
    "cython>=3.2.4",
    "sigkernel",
    "geomloss>=0.2.6",
    "ruff>=0.14.11",
    "nbconvert>=7.16.6",
    "ipywidgets>=8.1.8",
]

[dependency-groups]

cli = [
    # CLI-specific dependencies with proper version constraints
    "typer>=0.19.2",
    "rich>=14.1.0",
    "pyyaml>=6.0.2",
    "typer>=0.19.1",
    "click>=8.3.0",
    "colorama>=0.4.6",
    "tabulate>=0.9.0",
    "inquirer>=3.4.1",
    "prompt-toolkit>=3.0.52",
    "psutil>=7.1.0",
    "py-cpuinfo>=9.0.0",
    "questionary>=2.1.1",
    "textual>=6.1.0",
    "alive-progress>=3.3.0",
]
docs = [
    "sphinx>=4.0,<8.2",
    "sphinx_rtd_theme>=3.0.2",
    "myst_parser>=4.0.1",
    "nbsphinx>=0.9.7",
    "nbsphinx_link>=1.3.0",
    "sphinx_fontawesome>=0.0.6",
    "sphinx-autobuild>=2025.8.25",
    "recommonmark>=0.7.1",
    "sphinx_markdown_tables>=0.0.17",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=7.0.0",
    "black>=25.9.0",
    "flake8>=7.0.0",
    "isort>=6.0.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "flake8-docstrings>=1.7.0",
    "flake8-bugbear>=24.0.0",
    "bandit>=1.7.0",
    "pymarkdownlnt>=0.9.32",
    "torchtyping>=0.1.5",
]

logging = [
    # Optional logging backends used by LoggerConfig adapters. Install as:
    #   pip install .[logging]  or `pip install -e .[logging]` for editable installs
    "mlflow>=3.4.0",
    "neptune-client>=1.14.0",
    "comet-ml>=3.53.0",
]

[project.scripts]
new-ltpp = "scripts.cli:main"

[tool.uv.extra-build-dependencies]
git = ["Cython"]

[tool.uv.sources]
sigkernel = { git = "https://github.com/crispitagorico/sigkernel.git" }

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["new_ltpp", "scripts"]
exclude = ["tests*", "*.tests*", "*.tests"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.txt", "*.md"]

# Configuration for development tools
[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["new_ltpp"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["new_ltpp"]
omit = ["*/tests/*", "*/test_*", "setup.py", "version.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "pytorch_lightning.*",
    "wandb.*",
    "dvc.*",
    "hydra.*",
    "omegaconf.*",
    "torchmetrics.*",
    "fastdtw.*",
    "datasets.*",
    "polars.*",
]
ignore_missing_imports = true

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203", # whitespace before ':'
    "E501", # line too long (handled by black)
    "W503", # line break before binary operator
    "W504", # line break after binary operator
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".env",
    "venv",
    "env",
    "build",
    "dist",
    "*.egg-info",
    ".mypy_cache",
    ".pytest_cache",
    "docs/_build",
    "node_modules",
]
per-file-ignores = [
    "__init__.py:F401", # imported but unused
    "tests/*:D",        # no docstrings in tests
]
max-complexity = 10
